add defensive code to base index servie

+++ INSS.EIIR.AzureSearch/Services/BaseIndexService.cs
@@ -4,6 +4,7 @@ using Azure.Search.Documents.Indexes;
 using INSS.EIIR.Interfaces.SearchIndexer;
 using Azure.Search.Documents.Models;
 using INSS.EIIR.Models.IndexModels;
+using Azure;
 
 namespace INSS.EIIR.AzureSearch.Services;
 
@@ -22,19 +23,38 @@ public abstract class BaseIndexService<T> : IIndexService
     [ExcludeFromCodeCoverage]
     public async Task CreateIndexAsync()
     {
+
         var fieldBuilder = new FieldBuilder();
         var searchFields = fieldBuilder.Build(typeof(T));
 
         var definition = new SearchIndex(IndexName, searchFields);
 
-        await _searchClient.CreateOrUpdateIndexAsync(definition);
+        try
+        {
+            await _searchClient.CreateOrUpdateIndexAsync(definition);
+        }
+        catch (Exception)
+        {
+            Console.WriteLine($"Failed to create index {IndexName}");
+        }
     }
 
     [ExcludeFromCodeCoverage]
     public async Task DeleteIndexAsync()
     {
-        var index = await _searchClient.GetIndexAsync(IndexName);
-        await _searchClient.DeleteIndexAsync(index);
+        try
+        {
+            var index = await _searchClient.GetIndexAsync(IndexName);
+
+            if (index != null)
+            {
+                await _searchClient.DeleteIndexAsync(index);
+            }
+        }
+        catch (Exception)
+        {
+            Console.WriteLine($"Failed to delete index {IndexName}");
+        }
     }
 
     public abstract Task PopulateIndexAsync();
